+++ batcher/src/gui/widgets/dimension_box.py
@@ -123,6 +123,12 @@ class DimensionBox(Gtk.Box):
 
     self._update_tooltip()
 
+  def get_digits(self):
+    return self._spin_button.get_digits()
+
+  def set_digits(self, digits):
+    self._spin_button.set_digits(digits)
+
   def _init_gui(self):
     self.set_orientation(Gtk.Orientation.HORIZONTAL)
     self.set_spacing(self._widget_spacing)